Update for F19
Tidied things up for F19. This included adding some text that tries to
clarify the need for processing binary data, and how to use `flush()`.

I also removed all references to the port number, as we should be giving
them that in something like Canvas instead of cooking it into the lab.

Closes #1